requirements.txt
PyQt5>=5.15.0
PyQt5-sip>=12.0.0
aiohttp>=3.8.0
requests>=2.28.0
pycryptodome>=3.15.0
ffmpeg-python>=0.2.0
asyncio-throttle>=1.0.0
